# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Anatolia B.I.S. API",
    "version": "1.0.0",
    "description": "Machine-readable companion to API.md. See API.md for full prose documentation (request/response shapes, rate limits, error codes) — this file exists primarily so server/tests/openapi_drift.rs can catch a route that was added to, or removed from, server/src/routes/mod.rs without a matching documentation update. Kept intentionally lightweight rather than exhaustively typed."
  },
  "paths": {
...
    "/api/v1/audit": {
      "get": {
        "summary": "Paginated, filtered audit trail (AUDITOR/SECURITY_ADMIN/SYSTEM_ADMIN)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/api/v1/audit/integrity": {
      "get": {
        "summary": "Recomputes and verifies the audit-event hash chain",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK — { eventsChecked, intact, breaks }"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}